requirements.txt
asttokens==3.0.0
attrs==25.3.0
apache-airflow-core==3.0.1
colorama==0.4.6
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
executing==2.2.0
fastjsonschema==2.21.1
greenlet>=1.1.3,<3.2
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib==1.5.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
narwhals==1.38.2
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.2.5
packaging==25.0
pandas==2.2.3
parso==0.8.4
platformdirs==4.3.8
plotly==6.0.1
prompt_toolkit==3.0.51
psutil==7.0.0
psycopg2-binary==2.9.9
pure_eval==0.2.3
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
rpds-py==0.24.0
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
SQLAlchemy==1.4.52
stack-data==0.6.3
threadpoolctl==3.6.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
wcwidth==0.2.13
